deny_unknown_fields:false

+++ core/lib/dal/src/consensus_dal.rs
@@ -6,7 +6,6 @@ use zksync_db_connection::{
     error::{DalError, DalResult, SqlxContext},
     instrument::{InstrumentExt, Instrumented},
 };
-use zksync_protobuf::ProtoRepr as _;
 use zksync_types::L2BlockNumber;
 
 pub use crate::consensus::{proto, AttestationStatus, GlobalConfig, Payload};
@@ -55,14 +54,14 @@ impl ConsensusDal<'_, '_> {
         else {
             return Ok(None);
         };
+        let d = zksync_protobuf::serde::Deserialize {
+            deny_unknown_fields: true,
+        };
         if let Some(global_config) = row.global_config {
-            return Ok(Some(
-                zksync_protobuf::serde::deserialize(&global_config).context("global_config")?,
-            ));
+            return Ok(Some(d.proto_fmt(&global_config).context("global_config")?));
         }
         if let Some(genesis) = row.genesis {
-            let genesis: validator::Genesis =
-                zksync_protobuf::serde::deserialize(&genesis).context("genesis")?;
+            let genesis: validator::Genesis = d.proto_fmt(&genesis).context("genesis")?;
             return Ok(Some(GlobalConfig {
                 genesis,
                 registry_address: None,
@@ -80,9 +79,10 @@ impl ConsensusDal<'_, '_> {
     /// Resets the stored consensus state otherwise and purges all certificates.
     pub async fn try_update_global_config(&mut self, want: &GlobalConfig) -> anyhow::Result<()> {
         let mut txn = self.storage.start_transaction().await?;
-        if let Some(got) = txn.consensus_dal().global_config().await? {
-            // Exit if the genesis didn't change.
-            if &got == want {
+        let got = txn.consensus_dal().global_config().await?;
+        if let Some(got) = &got {
+            // Exit if the global config didn't change.
+            if got == want {
                 return Ok(());
             }
             anyhow::ensure!(
@@ -92,39 +92,40 @@ impl ConsensusDal<'_, '_> {
                 want.genesis.chain_id,
             );
             anyhow::ensure!(
-                got.genesis.fork_number < want.genesis.fork_number,
+                got.genesis.fork_number <= want.genesis.fork_number,
                 "transition to a past fork is not allowed: old = {:?}, new = {:?}",
                 got.genesis.fork_number,
                 want.genesis.fork_number,
             );
             want.genesis.verify().context("genesis.verify()")?;
         }
-        let genesis =
-            zksync_protobuf::serde::serialize(&want.genesis, serde_json::value::Serializer)
-                .unwrap();
-        let global_config =
-            zksync_protobuf::serde::serialize(want, serde_json::value::Serializer).unwrap();
-        let state = zksync_protobuf::serde::serialize(
-            &ReplicaState::default(),
-            serde_json::value::Serializer,
-        )
-        .unwrap();
-        sqlx::query!(
-            r#"
-            DELETE FROM l1_batches_consensus
-            "#
-        )
-        .instrument("try_update_genesis#DELETE FROM l1_batches_consensus")
-        .execute(&mut txn)
-        .await?;
-        sqlx::query!(
-            r#"
-            DELETE FROM miniblocks_consensus
-            "#
-        )
-        .instrument("try_update_genesis#DELETE FROM miniblock_consensus")
-        .execute(&mut txn)
-        .await?;
+        let s = zksync_protobuf::serde::Serialize;
+        let genesis = s
+            .proto_fmt(&want.genesis, serde_json::value::Serializer)
+            .unwrap();
+        let global_config = s.proto_fmt(want, serde_json::value::Serializer).unwrap();
+        let state = s
+            .proto_fmt(&ReplicaState::default(), serde_json::value::Serializer)
+            .unwrap();
+        if got.as_ref().map(|c| &c.genesis) != Some(&want.genesis) {
+            tracing::info!("consensus genesis changed, wiping consensus certificates");
+            sqlx::query!(
+                r#"
+                DELETE FROM l1_batches_consensus
+                "#
+            )
+            .instrument("try_update_genesis#DELETE FROM l1_batches_consensus")
+            .execute(&mut txn)
+            .await?;
+            sqlx::query!(
+                r#"
+                DELETE FROM miniblocks_consensus
+                "#
+            )
+            .instrument("try_update_genesis#DELETE FROM miniblock_consensus")
+            .execute(&mut txn)
+            .await?;
+        }
         sqlx::query!(
             r#"
             DELETE FROM consensus_replica_state
@@ -204,7 +205,13 @@ impl ConsensusDal<'_, '_> {
                 fake_key
             "#
         )
-        .try_map(|row| zksync_protobuf::serde::deserialize(row.state).decode_column("state"))
+        .try_map(|row| {
+            zksync_protobuf::serde::Deserialize {
+                deny_unknown_fields: true,
+            }
+            .proto_fmt(row.state)
+            .decode_column("state")
+        })
         .instrument("replica_state")
         .fetch_one(self.storage)
         .await
@@ -212,8 +219,9 @@ impl ConsensusDal<'_, '_> {
 
     /// Sets the current BFT replica state.
     pub async fn set_replica_state(&mut self, state: &ReplicaState) -> DalResult<()> {
-        let state_json =
-            zksync_protobuf::serde::serialize(state, serde_json::value::Serializer).unwrap();
+        let state_json = zksync_protobuf::serde::Serialize
+            .proto_fmt(state, serde_json::value::Serializer)
+            .unwrap();
         sqlx::query!(
             r#"
             UPDATE consensus_replica_state
@@ -303,7 +311,12 @@ impl ConsensusDal<'_, '_> {
         Ok(BlockStoreState {
             first: start,
             last: row
-                .map(|row| zksync_protobuf::serde::deserialize(row.certificate))
+                .map(|row| {
+                    zksync_protobuf::serde::Deserialize {
+                        deny_unknown_fields: true,
+                    }
+                    .proto_fmt(row.certificate)
+                })
                 .transpose()?,
         })
     }
@@ -331,7 +344,12 @@ impl ConsensusDal<'_, '_> {
         else {
             return Ok(None);
         };
-        Ok(Some(zksync_protobuf::serde::deserialize(row.certificate)?))
+        Ok(Some(
+            zksync_protobuf::serde::Deserialize {
+                deny_unknown_fields: true,
+            }
+            .proto_fmt(row.certificate)?,
+        ))
     }
 
     /// Fetches the attester certificate for the L1 batch with the given `batch_number`.
@@ -357,7 +375,12 @@ impl ConsensusDal<'_, '_> {
         else {
             return Ok(None);
         };
-        Ok(Some(zksync_protobuf::serde::deserialize(row.certificate)?))
+        Ok(Some(
+            zksync_protobuf::serde::Deserialize {
+                deny_unknown_fields: true,
+            }
+            .proto_fmt(row.certificate)?,
+        ))
     }
 
     /// Fetches a range of L2 blocks from storage and converts them to `Payload`s.
@@ -433,7 +456,9 @@ impl ConsensusDal<'_, '_> {
                 ($1, $2)
             "#,
             i64::try_from(header.number.0).context("overflow")?,
-            zksync_protobuf::serde::serialize(cert, serde_json::value::Serializer).unwrap(),
+            zksync_protobuf::serde::Serialize
+                .proto_fmt(cert, serde_json::value::Serializer)
+                .unwrap(),
         )
         .instrument("insert_block_certificate")
         .report_latency()
@@ -448,10 +473,9 @@ impl ConsensusDal<'_, '_> {
         number: attester::BatchNumber,
         committee: &attester::Committee,
     ) -> anyhow::Result<()> {
-        let committee = proto::AttesterCommittee::build(committee);
-        let committee =
-            zksync_protobuf::serde::serialize_proto(&committee, serde_json::value::Serializer)
-                .unwrap();
+        let committee = zksync_protobuf::serde::Serialize
+            .proto_repr::<proto::AttesterCommittee, _>(&committee, serde_json::value::Serializer)
+            .unwrap();
         sqlx::query!(
             r#"
             INSERT INTO
@@ -498,10 +522,11 @@ impl ConsensusDal<'_, '_> {
         else {
             return Ok(None);
         };
-        let raw = zksync_protobuf::serde::deserialize_proto(&row.attesters)
-            .context("deserialize_proto()")?;
         Ok(Some(
-            proto::AttesterCommittee::read(&raw).context("read()")?,
+            zksync_protobuf::serde::Deserialize {
+                deny_unknown_fields: true,
+            }
+            .proto_repr::<proto::AttesterCommittee, _>(row.attesters)?,
         ))
     }
 
@@ -535,7 +560,9 @@ impl ConsensusDal<'_, '_> {
             "#,
             i64::try_from(cert.message.number.0).context("overflow")?,
             // Unwrap is ok, because serialization should always succeed.
-            zksync_protobuf::serde::serialize(cert, serde_json::value::Serializer).unwrap(),
+            zksync_protobuf::serde::Serialize
+                .proto_fmt(cert, serde_json::value::Serializer)
+                .unwrap(),
         )
         .instrument("insert_batch_certificate")
         .report_latency()
